move nodeaffinity helpers to component-helpers package
